api/requirements.txt
Django==1.9.5
django-cors-headers==1.1.0
django-gcm==1.2.0
django-tastypie==0.13.3
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
geopy==1.11.0
Pillow==3.2.0
psycopg2==2.6.1
PyJWT==1.4.0
pyrcws==0.2.5
python-dateutil==2.5.3
python-mimeparse==1.5.1
pytz==2016.4
requests==2.9.1
six==1.10.0
suds==0.4
wheel==0.24.0
